requirements.txt
django==4.2
python-decouple==3.6
django-multi-token==0.1.5
django-filter==22.1
django-model-utils==5.0.0
celery==5.3.1
django-redis==5.3.0
django-celery-beat==2.5.0
aioredis==2.0.1
phonenumbers==8.12.50
django-phonenumber-field==6.2.0
djangorestframework==3.14.0
drf-writable-nested==0.6.4
psycopg2-binary==2.9.10
pillow==11.3.0
ipython==9.4.0